refactor HtmlWebpackPlugin usage across develop and build webpack configs
