+++ packages/cli/src/config/webpack.config.common.js
@@ -1,6 +1,5 @@
 const CopyWebpackPlugin = require('copy-webpack-plugin');
 const fs = require('fs');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
 const path = require('path');
 const webpack = require('webpack');
 
@@ -51,18 +50,6 @@ module.exports = ({ config, context }) => {
       }
     }
   ];
-
-  // gets Index Hooks to pass as options to HtmlWebpackPlugin
-  const customOptions = Object.assign({}, ...config.plugins
-    .filter((plugin) => plugin.type === 'index')
-    .map((plugin) => plugin.provider({ config, context }))
-    .filter((providerResult) => {
-      return Object.keys(providerResult).map((key) => {
-        if (key !== 'type') {
-          return providerResult[key];
-        }
-      });
-    }));
 
   // utilizes webpack plugins passed in directly by the user
   const customWebpackPlugins = config.plugins
@@ -124,13 +111,6 @@ module.exports = ({ config, context }) => {
           resource.request = resource.request.replace(/^\.\//, context.pagesDir);
         }
       ),
-
-      new HtmlWebpackPlugin({
-        filename: path.join(context.publicDir, context.indexPageTemplate),
-        template: path.join(context.scratchDir, context.indexPageTemplate),
-        chunksSortMode: 'dependency',
-        ...customOptions
-      }),
 
       ...customWebpackPlugins
     ]
